Fix ft::set insertion through red black tree
There were little mistakes in the code. Some of the main ones: (1)
the way I find the aunt of a node was wrong and (2) in the case where
the aunt is a black node I change the color of the wrong nodes and
perform rotation on the wrong node (in the case of L R and R L).

I also added some kind of small system for debugging that helped
me debug and check that the red black tree is being constructed
properly.