Adjust the nill node problem in rb_delete()
The main problem as far as i understand was that I didn't
check for nill nodes in the case of the deleted node being
root and thus, I leaked. So, I fixed that. Also, instead of
creating a nill node every time I delete, I just allocate
a nill node in the constructor and deallocate at destructor.