Differentiate between allocator_type and node_allocator
I faced a problem since I used allocator_type to be my node allocator.
This was an issue since in get_allocator() function, the prototype
specifies that the return type has to be allocator_type.. this would
(if it compiles, it doesn't) give the user access to a private class..
not to mention it wouldn't meet the requirements of mimicking std::map's
and std::set's behavior.