Fix vector's assign(), insert(), and erase()
There were issues when I constructed a pointer that had an existing
constructed object since I didn't call destroy on that pointer. Another
issue is that I can't call destroy always before constructing a pointer
as it could be that there is an absence of a constructed object. My
solution is a safeConstruct() that will check if the pointer is between
_start and _end and if so, I destroy before constructing. Otherwise,
I directly construct.